Refactored `TabScreen` and `ProjectSelector`
